requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
crispy-bootstrap5==2023.10
Django==4.2.8
django-bootstrap5==0.6.0
django-crispy-forms==2.1
djangorestframework==3.13.1
gunicorn==21.2.0
packaging==23.2
pytz==2023.3.post1
soupsieve==2.5
sqlparse==0.4.4
typing-extensions==4.9.0
whitenoise==5.3.0
